requirements.txt
boto3==1.34.84
google-cloud-storage==2.16.0
tenacity==8.2.3
